Move `io_*` functions to `base/io.cpp/h`
