+++ src/test/io_test.cpp
@@ -1,6 +1,9 @@
 #include "test.h"
 
-#include <base/system.h>
+#include <base/fs.h>
+#include <base/io.h>
+#include <base/mem.h>
+#include <base/str.h>
 
 #include <gtest/gtest.h>
